+++ admin/tabs/general/ehcache/EhCachePanel.js
@@ -5,7 +5,7 @@
 * Copyright © 2018 Extremely Heavy Industries Inc.
 */
 import {Component} from 'react';
-import {HoistComponent} from '@xh/hoist/core';
+import {HoistComponent, LoadSupport} from '@xh/hoist/core';
 import {grid} from '@xh/hoist/cmp/grid';
 import {filler} from '@xh/hoist/cmp/layout';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
@@ -17,6 +17,7 @@ import {Icon} from '@xh/hoist/icon';
 import {EhCacheModel} from './EhCacheModel';
 
 @HoistComponent
+@LoadSupport
 export class EhCachePanel extends Component {
 
     model = new EhCacheModel();
@@ -38,19 +39,11 @@ export class EhCachePanel extends Component {
             button({
                 icon: Icon.sync(),
                 text: 'Clear All',
-                onClick: this.onClearAllClick
+                onClick: () => model.clearAll()
             }),
             filler(),
             storeCountLabel({gridModel, unit: 'cache'}),
             storeFilterField({gridModel})
         );
-    }
-
-    onClearAllClick = () => {
-        this.model.clearAll();
-    }
-
-    async loadAsync() {
-        return this.model.loadAsync();
     }
 }