+++ admin/tabs/preferences/UserPreferencePanel.js
@@ -5,11 +5,12 @@
  * Copyright © 2018 Extremely Heavy Industries Inc.
  */
 import {Component} from 'react';
-import {HoistComponent} from '@xh/hoist/core';
+import {HoistComponent, LoadSupport} from '@xh/hoist/core';
 import {restGrid, RestGridModel, RestStore} from '@xh/hoist/desktop/cmp/rest';
 import {usernameCol} from '@xh/hoist/admin/columns';
 
 @HoistComponent
+@LoadSupport
 export class UserPreferencePanel extends Component {
 
     model = new RestGridModel({
@@ -80,9 +81,5 @@ export class UserPreferencePanel extends Component {
 
     render() {
         return restGrid({model: this.model});
-    }
-
-    async loadAsync() {
-        return this.model.loadAsync();
     }
 }